services/fitnessclub/requirements.txt
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
zipp==3.8.0
gunicorn==20.1.0
flask_cors==3.0.10
Werkzeug==2.2.2
python-dotenv==0.21.1
jsonify>=0.5
requests>=2.27
Flask-Cors==3.0.10
datetimerange>=2.1.0
Flask>=2.2
requests>=2,<3
ms_identity_python[flask] @ https://github.com/azure-samples/ms-identity-python/archive/refs/heads/0.9.zip
azure-storage-blob>=12.25.0
